feat: add pages() method to User model
This commit adds a `pages()` method to the `User` model.

The `pages()` method returns a `HasMany` relationship to the `Page` model, where the `creator_id`
column in the `pages` table references the `id` column in the `users` table. This allows the `isReferenced()` method to correctly identify if a user is referenced through pages.