refactor: remove team features and update factories
This commit refactors the application by removing the team features and updating the factories.

The team feature was disabled by commenting it out in the `config/jetstream.php` file.

The MediaFactory was updated to use the User factory instead of generating a random ulid for the creator_id. It also sets default values for disk to `public` and updates some faker calls.

The UserFactory's withPersonalTeam method was removed.

Tests that were using `withPersonalTeam` have been updated to not use it.