feat: update language switcher and fix x-persist issue
This commit updates the language switcher in the admin panel, allowing users to change the application's locale.

The language switcher is implemented as a dropdown menu in the navigation bar and includes options for Bahasa Indonesia and English. Each language option is enhanced witha checkmark icon to indicate the selected language and active UI styling to visually highlight the currently active language. The `language-switcher-item` blade component was created to reduce code duplication for each language option.

This commit also fixes an issue where `x-persist` was interfering with the language switcher's functionality. A workaround was implemented by rendering a Blade component before the global search to ensure the `x-persist` context is properly terminated, resolving the conflict.