feat: add device detection middleware
This commit introduces a new middleware `SetDeviceFromHeader` that detects the device type (Android) and the value of the `x-requested-with` header from the request headers.

The device information is stored in the `App\Utils\Device` class, which provides static properties to check if the request is coming from an Android device or has a specific value for the 'x-requested-with' header.

This information can be used later in the application to customize the response based on the device type.

The middleware is added to the global middleware stack in `bootstrap/app.php`.