refactor: remove phpstan ignore
This commit removes the `@phpstan-ignore-next-line` annotation from the `AppServiceProvider.php` file.

The annotation was no longer necessary as the code now correctly handles the case where `$locales` is null. The previous implementation was throwing an exception if `$locales` was null, ensuring that the code would not proceed with a null value. Therefore, the `@phpstan-ignore-next-line` was redundant and has been removed to improve code clarity.