feat: add database notifications to navigation menu
This commit introduces database notifications to the navigation menu. It adds a new Livewire component that displays a bell icon with a badge indicating the number of unread notifications. Clicking the icon likely opens a modal or dropdown with the list of notifications (though the precise behavior of the `database-notifications` Livewire component isn't defined in this diff). It also triggers the DatabaseNotifications event to mark the notification as read.