test: add test to check correct records are listed in export resource
This commit adds a test case to ensure that the ExportResource only lists the exports belonging to the currently logged-in user (either created by them or associated with their user ID). The test creates several exports with different user and creator IDs, then asserts that the Livewire component only displays the exports associated with the currently authenticated user.